+++ apps/spruce/src/pages/preferences/preferencesTabs/cliTab/AuthenticationCard.tsx
@@ -6,13 +6,11 @@ import { CardSkeleton } from "@leafygreen-ui/skeleton-loader";
 import { Subtitle } from "@leafygreen-ui/typography";
 import get from "lodash/get";
 import { size } from "@evg-ui/lib/constants/tokens";
+import { postAndHandle } from "@evg-ui/lib/utils/request/post";
 import { usePreferencesAnalytics } from "analytics";
 import { SettingsCard } from "components/SettingsCard";
 import { UserConfigQuery, UserConfigQueryVariables } from "gql/generated/types";
 import { USER_CONFIG } from "gql/queries";
-import { request } from "utils";
-
-const { post } = request;
 
 export const AuthenticationCard = () => {
   const { data, loading, refetch } = useQuery<
@@ -34,7 +32,7 @@ ui_server_host: "${config?.ui_server_host}"
   const resetKey = async (e) => {
     e.preventDefault();
     sendEvent({ name: "Clicked reset API key" });
-    await post(`/settings/newkey`, {});
+    await postAndHandle(`${getEvergreenUrl()}/settings/newkey`, {});
     refetch();
   };
   // @ts-expect-error: FIXME. This comment was added by an automated script.
