+++ packages/lib/src/utils/request/request.test.ts
@@ -1,3 +1,4 @@
+import { postAndHandle } from "./post";
 import { fetchWithRetry } from ".";
 
 describe("request utils", () => {
@@ -76,6 +77,53 @@ describe("request utils", () => {
         message: "Internal Server Error",
       });
       expect(global.fetch).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe("postAndHandle", () => {
+    afterEach(() => {
+      vi.clearAllMocks();
+    });
+
+    it("should make a POST request and return the response for a successful request", async () => {
+      const url = "/api/resource";
+      const body = { key: "value" };
+      const fetchMock = vi.fn().mockResolvedValue({
+        ok: true,
+      });
+
+      vi.spyOn(global, "fetch").mockImplementation(fetchMock);
+
+      const response = await postAndHandle(url, body);
+
+      expect(response).toStrictEqual({ ok: true });
+      expect(fetchMock).toHaveBeenCalledWith("/api/resource", {
+        method: "POST",
+        body: JSON.stringify(body),
+        credentials: "include",
+      });
+    });
+
+    it("should handle and report an error for a failed request", async () => {
+      const url = "/api/resource";
+      const body = { key: "value" };
+      const fetchMock = vi.fn().mockResolvedValue({
+        ok: false,
+        status: 500,
+        statusText: "Internal Server Error",
+      });
+      const errorReportingMock = vi.fn();
+      vi.spyOn(console, "error").mockImplementation(errorReportingMock);
+      vi.spyOn(global, "fetch").mockImplementation(fetchMock);
+
+      await postAndHandle(url, body);
+
+      expect(fetchMock).toHaveBeenCalledWith("/api/resource", {
+        method: "POST",
+        body: JSON.stringify(body),
+        credentials: "include",
+      });
+      expect(errorReportingMock).toHaveBeenCalledTimes(1);
     });
   });
 });
